pyproject.toml
[project]
name = "alpha-guide"
version = "0.1.0"
description = "gpt-like-a-dev"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115.12",
    "jupyter>=1.1.1",
    "openai",
    "pydantic>=2.11.4",
    "uvicorn>=0.34.2",
    "langchain>=1.0.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.3.0",
    "langgraph-checkpoint>=2.0",
    "langgraph-checkpoint-redis>=0.0.1",
    "python-multipart>=0.0.18",
    "langchain-text-splitters>=0.3.0",
    "langsmith>=0.2.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.13.2",
    "langchain-qdrant>=0.2.0",
    "langchain-classic>=0.1.0",
    #"ragas==0.2.10",
    #"nltk==3.8.1",
    #"openevals>=0.1.3",
    #"pypdf>=5.2.0",
    #"langchain-tavily>=0.1.0",
    "tavily-python>=0.7.23",
    "mcp[cli]>=1.12.0",
    "tradingview-screener>=0.6.4",
    "tradingview-ta>=3.3.0",
]
